refactor: Processor 분리
