+++ Moyoy-Batch/src/main/java/com/moyoy/batch/helper/GithubCommitStatCalculator.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 import org.springframework.stereotype.Component;
 
-import com.moyoy.batch.dto.GithubRepoCommitStats;
+import com.moyoy.batch.job.processor.dto.GithubRepoCommitStats;
 
 import com.moyoy.domain.ranking.GithubCommitStats;
