+++ Moyoy-Batch/src/main/java/com/moyoy/batch/job/processor/dto/GithubRepoCommitStats.java
@@ -1,4 +1,4 @@
-package com.moyoy.batch.dto;
+package com.moyoy.batch.job.processor.dto;
 
 import java.util.List;
