+++ Moyoy-Batch/src/main/java/com/moyoy/batch/job/processor/dto/UserRankResult.java
@@ -1,4 +1,4 @@
-package com.moyoy.batch.dto;
+package com.moyoy.batch.job.processor.dto;
 
 import com.moyoy.domain.ranking.dto.RankingCalculatorResult;
